refactoring -moved data to media -moved assets to one dir -removed unused font from vendor dir
