+++ photobooth/assets/systemservice/boothupload.sh
